rename github to github_utils to avoid import circular dependency problems
